feat: Hook the Map component to snowflakeDB
The map component was using dummy data to display the different markers and information.
As we are closing to the go-live moment, the map needed to talk to the snowflakeDB
and get its data from there.

Not all information is in the DB just yet but the attempt was to use as much as possible
to reduce work in a couple days when the DB is updated again.