Use translations for the main_page
The current page has a lot of translated text in it. This allows for dynamic updates but
the current table structure does not allow for more than grammar fixes.

Besides that we are now using all the translations that are available in the table
if there are any values that are not available then there are fallback values that
can set the data to what seems logical. If at any point the data becomes available in
the database then the fallback should use those values unless the column name changes.

The resulting `translateOrFallback` function is not pretty by any means, but it is a solution
that works for our immediate use case and allows us to ignore most of the translations
on the page for now.