fix: Update MapHeaderContainer to handle new data from Snowflake
Someone updated the SnowflakeDB during the night and this update changed a lot
in regard to the Map data that we had available. The data changed so much
that many of the things that were previously working were now broken.

This update fixes all the bugs that were caused by that update and hopefully
made the code a bit more resilient against potential breakages.