fix: filter Map markers by country
This is a bug that was introduced while hooking up the SnowflakeDB to the
data.

If the dropdown is used to filter based on country, then we only want to show
markers for that specific country. If the filter is used to further select
"program types" then the markers should be filtered deeper based on program type.