fix: Disable map zoom and pan
I was not able to reproduce this behavior anymore, but it was reported that
zooming and panning the map would sometimes cause weird jumping behavior.

Simply disabling all the scrolling and panning from a simple object means we
toggle map features on and off based on the current state of testing.
If during testing no issues are found with a certain feature we can always turn
it back on, if there are issues we now have a simple way to turn it off.