fix: year dropdown only shows years
The year dropdown should not have any "all" filter. It should
by default select the most current year and all other years after
that.

fix: restructure a part of main_page

As a prep for a refactor, I decided to first get a better
overview what is exactly happening on the main_page and slightly restructure
the calls that are being done there.

fix: resolve error onLoad every page

Every page would show an error that you cannot pass a `string` when a function is expected.
In older versions of HTML we had to jump through hoops to ensure our `onLoad` function
would be properly called, but in recent versions, and especially when using JSX
we can simply pass a function to be executed

refactor: simplify useMapMarkers

Destructuring the array makes the code a bit easier to understand

fix: add typescript to fix `yarn lint`

I was banging my head against these stupid ESLint errors that were showing and breaking my editor…
and I think the problem is closely related to the build problems that were showing

We have several `tsx` files within the project, but we do not have typescript as a dependency.
This will freak out ESLint as starting from next 13, typescript is the default.
I guess this completely went under the radar because `next build` will actually use `tsc` under the hood for typescript files.

Adding typescript as dependency made `yarn lint` work again for me, but it now
does show a bunch of warnings/errors plus we need a tsconfig file that sets `esModuleInterop`.

Of course, not a 100% certain that this is where the build problems come from, but nonetheless
its always great to have ESLint check for potential problems within during coding

```╰─λ find . -type f -name "*.tsx"
./Tabs/Tabs.stories.tsx
./AccordionGroup/AccordionGroup.stories.tsx
./Dropdown/Dropdown.tsx
./Dropdown/Dropdown.stories.tsx
./Charts/BarChart/BarChart.stories.tsx
./Charts/BarChart/BarChart.tsx
./Charts/StackedBarChart/StackedBarChart.tsx
./Charts/StackedBarChart/StackedBarChart.stories.tsx
./Charts/DoughnutChart/DoughnutChart.stories.tsx
./LogoBlock/LogoBlock.stories.tsx
./Button/Button.stories.tsx
./CtaBlock/CtaBlock.stories.tsx
./AccordionItem/AccordionItem.stories.tsx
./EmergencyAlert/ExclamationMark.tsx```

refactor: change the way data flows into the map

The map was initially developed to be super layered where there were a lot of things passed back
and forth and it became incredibly hard to debug where problems came from.

Next to a bunch of changes and fixes this removes the `MapHeaderContainer` as it was not actually doing anything
useful. Initially it now feels that `MapChartHeader` is overloaded with code but it is doing the exact same thing
as was already previously happening.

As I will slowly move things around I hope it becomes more clear what the different moving parts do and what
their responsibilities are.

update countryCode to use value from snowflake