refactor: change the way data flows into the map
The map was initially developed to be super layered where there were a lot of things passed back
and forth and it became incredibly hard to debug where problems came from.

Next to a bunch of changes and fixes this removes the `MapHeaderContainer` as it was not actually doing anything
useful. Initially it now feels that `MapChartHeader` is overloaded with code but it is doing the exact same thing
as was already previously happening.

As I will slowly move things around I hope it becomes more clear what the different moving parts do and what
their responsibilities are.