fix: display map markers on country pages
The markers would never properly render on the country template, even though they should
and the code was almost identical as the main page. The main difference between the two
was that "Home" was using `MapChartHeader` while "Country" is directly using `MapChartCountry`.

After a lot of debugging and banging my head I realised that in 7955c3587bda0 I fixed a similar issue
for the main page map that was doing funky stuff. Solution there was to simply import the
`mapbox-gl.css` file to ensure everything was working as expected.

This fix is really nothing more than passing the correct coordinates and then adding the CSS file...