fix: filter program dropdown on country selection
The program dropdown should only show programs types that are available for a
country. If a program type is not available then we should filter it from the
dropdown and not show it at all.

This behavior was already intended but during one of the refactors it broke silently.
Instead of checking for the specific country AND selected year it checked for the
specific country OR the selected year and thus always returned true for all program types
even if they were not available for a country.